Refactor restore to new project components
