+++ apps/studio/data/projects/clone-status-query.ts
@@ -1,7 +1,10 @@
-import { get, handleError } from 'data/fetchers'
-import { projectKeys } from './keys'
 import { useQuery, UseQueryOptions } from '@tanstack/react-query'
+import { components } from 'api-types'
+import { get, handleError } from 'data/fetchers'
 import { ResponseError } from 'types'
+import { projectKeys } from './keys'
+
+export type CloneStatus = components['schemas']['ProjectClonedStatusResponse']
 
 export async function getCloneStatus(projectRef?: string) {
   if (!projectRef) throw new Error('Project ref is required')
