+++ packages/common/enabled-features/enabled-features.json
@@ -32,7 +32,7 @@
 
   "database:replication": true,
   "database:roles": true,
-  "database:restore_to_new_project": false,
+  "database:restore_to_new_project": true,
 
   "docs:auth_architecture": true,
   "docs:auth_configuration": true,
